Groups and group permissions (#122)
* Controller generator

* Adds user groups

* Implements group assignment

* Tried to make generators

* Implements user group views

* Updates many views

* Combines user and person forms

* Updates packages

* I need to eat something

* Working on Person form with useInertiaForm

* Spelling

* Adds yarn error log to gitignore

* Updates gems

* Adds login enabled toggle to User form

* Corrects form section hiding

* Adds full reload for vite upon routes change

* Yarn update

* Removes vitest/globasl from tsconfig

* Adds configurable local allow list to vite config

* Adds working generators

* Removes guard live reload

* Removes final flatpicker mentions

* Removes trustifi

* UseForm typings

* Minor stuff

* More work on the person form

* Removes jest, adds a test

* Rearranges settings page

* Implements routed settings page

* Trying things

* Adds accessibility labels

* Adds incrementing ticket numbers to tickets

* LDAPS save from new settings layout

* Adds group permissions form

* Adds tiptap and cleans up inputs

* Updates dynamic form implementation

* Adds pundit

* Removes cancancan

* Fixes factories

* Updates website schema

* More work on groups

* Permissions are saving

* Moves permission types to declaration file

* Cleans up gemfile

* Types

* Renames variables

* Fixes user group types

* Company admin switch

* Group permissions save

* Policies

* Starts to rename groups table

* Adds errorKey to form inputs

* Fixes model associations

* Fixes role migrations and seeds

* Groups load properly

* Fixes person tests

* Fixes person request test

* All tests pass

* Adds policies

* Adds authorize checks to controller actions

* Adds color picker to settings

* Reverts color on navigate away

* Can save appearance settings

* Color scheme saves and navigates properly

* Adds some tests

* Updates specs

* Tests pass

* downcases the assignable folder name

* OJ Serializers (#121)

* Adds OJ serializer

* Adds basic serializers

* Successfully exports all serializer types

* Tests Pass

* Removes blueprinter

+++ app/controllers/companies_controller.rb
@@ -2,11 +2,11 @@ class CompaniesController < ApplicationController
   include Searchable
 
   expose :companies, -> { search(current_user.companies, sortable_fields) }
-  expose :company, -> { current_user.companies.find_by_slug(params[:slug]) }
+  expose :company, id: ->{ params[:slug] }, scope: ->{ current_user.companies }, find_by: :slug
 
   # GET /companies
-  # GET /companies.json
   def index
+    authorize companies
     paginated_companies = companies.page(params[:page] || 1)
 
     render inertia: "Companies/Index", props: {
@@ -20,17 +20,19 @@ class CompaniesController < ApplicationController
 
   # GET /companies/:slug
   def show
+    authorize company
     if company.nil?
       render inertia: "Error", props: { status: 404 }
     else
       render inertia: "Companies/Show", props: {
-        company: -> { company.render(view: :associations) }
+        company: -> { company.render }
       }
     end
   end
 
   # GET /companies/new
   def new
+    authorize Company
     render inertia: "Companies/New", props: {
       company: Company.new.render(view: :new)
     }
@@ -38,6 +40,7 @@ class CompaniesController < ApplicationController
 
   # GET /companies/:slug/edit
   def edit
+    authorize company
     render inertia: "Companies/Edit", props: {
       company: -> { company.render(view: :edit) }
     }
@@ -45,6 +48,7 @@ class CompaniesController < ApplicationController
 
   # POST /companies
   def create
+    authorize Company
     if Company.new(company_params).save
       # Assign admin permissions to user creating the record
       current_user.add_role :admin, company
@@ -58,6 +62,7 @@ class CompaniesController < ApplicationController
 
   # PATCH/PUT /companies/:slug
   def update
+    authorize company
     if company.update(company_params)
       redirect_to company, notice: 'Company was successfully updated.'
     else
@@ -67,6 +72,7 @@ class CompaniesController < ApplicationController
 
   # DELETE /companies/:slug
   def destroy
+    authorize company
     company.destroy
     respond_to do |format|
       format.html { redirect_to companies_url, notice: 'Company was successfully destroyed.' }
