+++ app/controllers/items_controller.rb
@@ -3,14 +3,16 @@ class ItemsController < ApplicationController
   include Searchable
 
   expose :items, -> { search(@active_company.items.includes_associated, sortable_fields) }
-  expose :item
+  expose :item, scope: ->{ @active_company.items }, find: ->(id, scope){ scope.includes_associated.find(id) }
 
   # GET /item
   def index
+    authorize items
     paginated_items = items.page(params[:page] || 1)
 
     render inertia: "Items/Index", props: {
-      items: -> { paginated_items.render(view: :index) },
+      # items: -> { paginated_items.render(view: :index) },
+      items: -> { Items::IndexSerializer.many(paginated_items) },
       pagination: -> { {
         count: items.count,
         **pagination_data(paginated_items)
@@ -27,6 +29,7 @@ class ItemsController < ApplicationController
 
   # GET /item/:id
   def show
+    authorize item
     render inertia: "Items/Show", props: {
       item: -> { item.render(view: :show) }
     }
@@ -34,39 +37,43 @@ class ItemsController < ApplicationController
 
   # GET /item/new
   def new
+    authorize Item
     render inertia: "Items/New", props: {
       item: Item.new.render(view: :new),
-      models: -> { @active_company.models.find_by_category(:Item).render(view: :as_options) },
-      vendors: -> { @active_company.vendors.render(view: :as_options) },
-      locations: -> { @active_company.locations.render(view: :as_options) },
-      manufacturers: -> { @active_company.manufacturers.render(view: :as_options) },
-      categories: -> { @active_company.categories.find_by_type(:item).render(view: :as_options) }
+      models: -> { @active_company.models.find_by_category(:Item).render(view: :options) },
+      vendors: -> { @active_company.vendors.render(view: :options) },
+      locations: -> { @active_company.locations.render(view: :options) },
+      manufacturers: -> { @active_company.manufacturers.render(view: :options) },
+      categories: -> { @active_company.categories.find_by_type(:item).render(view: :options) }
     }
   end
 
   # GET /item/:id/edit
   def edit
+    authorize item
     render inertia: "Items/Edit", props: {
       item: item.render(view: :edit),
-      models: -> { @active_company.models.find_by_category(:Item).render(view: :as_options) },
-      vendors: -> { @active_company.vendors.render(view: :as_options) },
-      locations: -> { @active_company.locations.render(view: :as_options) },
-      manufacturers: -> { @active_company.manufacturers.render(view: :as_options) },
-      categories: -> { @active_company.categories.find_by_type(:item).render(view: :as_options) }
+      models: -> { @active_company.models.find_by_category(:Item).render(view: :options) },
+      vendors: -> { @active_company.vendors.render(view: :options) },
+      locations: -> { @active_company.locations.render(view: :options) },
+      manufacturers: -> { @active_company.manufacturers.render(view: :options) },
+      categories: -> { @active_company.categories.find_by_type(:item).render(view: :options) }
     }
   end
 
   # GET /item/:id/clone
   def clone
-    self.item = Item.find(params[:id]).dup
-    self.item.serial = nil
-    self.item.item_tag = nil
-    self.item
+    authorize item
+    cloned_item = item.dup
+    cloned_item.serial = nil
+    cloned_item.item_tag = nil
+
     render inertia: "Items/Clone"
   end
 
   # GET /item/:id/checkout
   def checkout
+    authorize item
     if item.assigned?
       redirect_to item, warning: 'Item is already checked out'
     else
@@ -75,15 +82,16 @@ class ItemsController < ApplicationController
       render inertia: "Items/Checkout", props: {
         item: item.render,
         assignment: assignment.render(view: :new),
-        people: -> { @active_company.people.select([:id, :first_name, :last_name, :location_id]).render(view: :as_options) },
-        items: -> { @active_company.items.select([:id, :name, :default_location_id]).render(view: :as_options) },
-        locations: -> { @active_company.locations.render(view: :as_options) },
+        people: -> { @active_company.people.select([:id, :first_name, :last_name, :location_id]).render(view: :options) },
+        items: -> { @active_company.items.select([:id, :name, :default_location_id]).render(view: :options) },
+        locations: -> { @active_company.locations.render(view: :options) },
       }
     end
   end
 
   # GET /item/:id/checkin
   def checkin
+    authorize item
     if item.assigned?
       assignment = item.assignment
       assignment.returned_at = Time.current
@@ -92,7 +100,7 @@ class ItemsController < ApplicationController
       render inertia: "Items/Checkin", props: {
         item: item.render,
         assignment: assignment.render,
-        locations: -> { @active_company.locations.render(view: :as_options) },
+        locations: -> { @active_company.locations.render(view: :options) },
         statuses: -> { StatusLabel.all.render } # TODO: Is this scoped to a Company?
       }
     else
@@ -102,6 +110,8 @@ class ItemsController < ApplicationController
 
   # POST /item
   def create
+    authorize Item
+    item = Item.new(item_params)
     item.company = @active_company
 
     if item.save
@@ -113,6 +123,7 @@ class ItemsController < ApplicationController
 
   # PATCH/PUT /item/:id
   def update
+    authorize item
     if item.update(item_params)
       redirect_to item, notice: 'Item was successfully updated'
     else
@@ -122,6 +133,7 @@ class ItemsController < ApplicationController
 
   # DELETE /item/:id
   def destroy
+    authorize item
     item.destroy
     redirect_to items_url, notice: 'Item was successfully destroyed.'
   end
